Changed Prediction Error
The error is being caused due to the updation mismatch between count and current_state's shape. The tensor moves with the updation of current_state and the previous tensor, however count only updates by one, this leads to an error where count may be 500 but the current_state's shape is > 512. The updation strictly should take place with accordance to the input tensor.